Refactor device-discovery
Separate it into 4 files for the 4 separate classes
